Check Interpretation Phi Argument

Write a textual interpretation of the values in Phi. This can be used to check if Phi has been correctly specified.

Usage

check_Phi(Phi)

Arguments

Phi

A matrix, with standardized autoregressive effects (on the diagonal) and cross-lagged effects (off-diagonal) in the population. Columns represent predictors and rows represent outcomes.

Value

No return value, called for side effects.

Examples

# Correctly specified `Phi`
Phi1 <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
check_Phi(Phi1)
#> According to this `Phi`, the lagged effects are:
#> • Autoregressive effect of A: 0.4
#> • Autoregressive effect of B: 0.3
#> • Cross-lagged effect of A -> B: 0.2
#> • Cross-lagged effect of B -> A: 0.1

# `Phi` with too large standardized effects
Phi2 <- matrix(c(.6, .5, .4, .7), ncol = 2, byrow = TRUE)
Phi2 <- check_Phi(Phi2)
#> According to this `Phi`, the lagged effects are:
#> • Autoregressive effect of A: 0.6
#> • Autoregressive effect of B: 0.7
#> • Cross-lagged effect of A -> B: 0.4
#> • Cross-lagged effect of B -> A: 0.5
#> 
#> However, lagged effects in `Phi` must specify a stationary process:
#> ℹ This is checked by testing if the eigenvalues of `Phi` lie within unit circle.
#> ✖ The eigenvalues of `Phi` are not within unit circle. Try out smaller lagged effects?