Check Interpretation of Lagged Effects

Write a textual interpretation of the values in lagged_effects. This can be used to check if lagged_effects has been correctly specified.

Usage

check_lagged_effects(lagged_effects = NULL, Phi = NULL, ...)

check_Phi(lagged_effects = NULL, Phi = NULL)

Arguments

lagged_effects

A matrix, with standardized autoregressive effects (on the diagonal) and cross-lagged effects (off-diagonal) in the population. Columns represent predictors and rows represent outcomes.

Phi

Alternative name for lagged_effects.

...

Additional arguments are not allowed.

Value

No return value, called for side effects.

Details

check_Phi() is retained for users who prefer the Phi notation.

Examples

# Correctly specified lagged effects
lagged_effects1 <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
check_lagged_effects(lagged_effects1)
#> According to `lagged_effects`, the lagged effects are:
#> • Autoregressive effect of A: 0.4
#> • Autoregressive effect of B: 0.3
#> • Cross-lagged effect of A -> B: 0.2
#> • Cross-lagged effect of B -> A: 0.1

# Lagged effects with too large standardized effects
lagged_effects2 <- matrix(c(.6, .5, .4, .7), ncol = 2, byrow = TRUE)
lagged_effects2 <- check_lagged_effects(lagged_effects2)
#> According to `lagged_effects`, the lagged effects are:
#> • Autoregressive effect of A: 0.6
#> • Autoregressive effect of B: 0.7
#> • Cross-lagged effect of A -> B: 0.4
#> • Cross-lagged effect of B -> A: 0.5
#> 
#> However, `lagged_effects` must specify a stationary process:
#> ℹ This is checked by testing if the eigenvalues of `lagged_effects` lie within unit circle.
#> ✖ The eigenvalues of `lagged_effects` are not within unit circle. Try out smaller lagged effects?
lagged_effects <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
check_Phi(lagged_effects)
#> According to `lagged_effects`, the lagged effects are:
#> • Autoregressive effect of A: 0.4
#> • Autoregressive effect of B: 0.3
#> • Cross-lagged effect of A -> B: 0.2
#> • Cross-lagged effect of B -> A: 0.1