Check Interpretation of Random-Intercept Loadings

Write a textual interpretation of the values in loadings. This can be used to check if time-varying random-intercept loadings have been correctly specified for lavaan data generation.

Usage

check_loadings(loadings = NULL, time_points = NULL, ...)

Arguments

loadings

A numeric vector or matrix specifying random-intercept loadings in the lavaan data-generating model. A vector of length time_points is applied to both variables. A matrix must have two rows, one for each variable, and one column per time point. The first loading must be 1, so later values are interpreted relative to the first occasion.

time_points

(optional) A single integer indicating the number of time points. If omitted, this is inferred from loadings.

...

Additional arguments are not allowed.

Value

No return value, called for side effects.

Examples

# Same random-intercept loadings for A and B
loadings1 <- c(1, .8, 1.1, .9)
check_loadings(loadings1)
#> According to `loadings`, the random-intercept loadings in the data-generating model are:
#> • RI_A loads on A1 and RI_B loads on B1 with 1.
#> • RI_A loads on A2 and RI_B loads on B2 with 0.8.
#> • RI_A loads on A3 and RI_B loads on B3 with 1.1.
#> • RI_A loads on A4 and RI_B loads on B4 with 0.9.

# Different random-intercept loadings for A and B
loadings2 <- matrix(c(1, .8, 1.1, .9, 1, 1.2, .7, 1), nrow = 2, byrow = TRUE)
check_loadings(loadings2, time_points = 4)
#> According to `loadings`, the random-intercept loadings in the data-generating model are:
#> • RI_A loads on A1 with 1.
#> • RI_A loads on A2 with 0.8.
#> • RI_A loads on A3 with 1.1.
#> • RI_A loads on A4 with 0.9.
#> 
#> • RI_B loads on B1 with 1.
#> • RI_B loads on B2 with 1.2.
#> • RI_B loads on B3 with 0.7.
#> • RI_B loads on B4 with 1.