Power analysis for the RI-CLPM

powRICLPM() performs a Monte Carlo power analysis for the random intercept cross-lagged panel model (RI-CLPM). It computes performance metrics for all RI-CLPM parameters (e.g., bias, mean square error, coverage, power, etc), and can perform power analysis across multiple experimental conditions simultaneously. Conditions can vary in terms of sample size, number of time points, and proportion of between-unit variance (i.e., ICC). See "Details" for information on the data simulation/estimation, internal naming conventions of parameters, and parallel execution options for speeding up the analysis.

Usage

powRICLPM(
  target_power,
  search_lower = NULL,
  search_upper = NULL,
  search_step = 20,
  sample_size = NULL,
  time_points,
  ICC,
  RI_cor,
  Phi,
  wSigma,
  skewness = 0,
  kurtosis = 0,
  alpha = 0.05,
  reps = 1000,
  seed = NA,
  save_path = NA,
  parameter = NULL,
  constraints = NULL,
  bounds = FALSE
)

Arguments

target_power

A numeric value between 0 and 1, denoting the targeted power level.

search_lower

A positive integer, denoting the lower bound of the range of sample sizes.

search_upper

A positive integer, denoting the upper bound of the range of sample sizes.

search_step

A positive integer, denoting the increment in sample size.

sample_size

An integer (vector) with elements at least larger than 20 (i.e., the number of free parameters in a basic bivariate RI-CLPM with 3 repeated measures), indicating the sample sizes at which to evaluate power. This argument can be specified as an alternative to the `search_` arguments to denote specific sample sizes, rather than an entire range.

time_points

An integer (vector) with elements larger than 3, indicating number of time points.

ICC

A `double` denoting the proportion of variance at the between-unit level.

RI_cor

A `double` denoting the correlation between random intercepts.

Phi

A matrix of standardized autoregressive and cross-lagged effects in the population. Columns represent predictors and rows represent outcomes.

wSigma

A correlation matrix for the within-unit components.

skewness

A numeric value denoting the skewness values for the observed variables. For more information, see simulateData.

kurtosis

A numeric value denoting the kurtosis values for the observed variables. For more information, see simulateData.

alpha

A `double` denoting the significance criterion (defaults to 0.05).

reps

A positive `integer` denoting the number of Monte Carlo replications to be used during simulations.

seed

An `integer` of length 1. If multiple cores are used, a seed of length 1 will be used to generate a full L'Ecuyer-CMRG seed for all cores. For more information, see furrr_options.

save_path

A character string naming the directory to save any (data) files to.

parameter

A character string denoting the parameter of primary interest. If specified, coef, summary, and plot output results specifically for this parameter.

constraints

A character string specifying the RI-CLPM parameters that should be constrained over time in the estimation model. "lagged" refers to both cross-lagged and autoregressive effects, "residuals" refers to the residual variances, "within" refers to both lagged effects and residual variances, and "stationarity" imposes stationarity constraints on the lagged effects and residual variances. By default, the lagged effects and residual variances are allowed to vary freely over time.

bounds

A logical denoting if bounded estimation should be used for the latent variable variances in the model. This can only be used when no constraints are imposed on the estimation model. See "Details". Bounded estimation is useful to avoid nonconvergence in small samples (< 100). Here, automatic wide bounds are used as advised by De Jonckere and Rosseel (2022). For more information, see optim.bounds in lavOptions.

Value

A list containing a "conditions" and "session" element. The "condition" element is again a list of experimental conditions, where each element is itself a list containing the input and output of the power analysis for a specific experimental condition. This includes:

sample_size: A sample size.

time_points: A number of time points.

ICC: A proportion of between-unit variance.

RI_var: Random intercept variance (based on ICC).

RI_cov: Random intercept covariance (based on ICC and RI_var).

reps: The number of Monte Carlo replication.

pop_synt: lavaan model syntax containing population values for data generation.

pop_tab: lavaan parameter table for data generation.

est_synt: lavaan model syntax for estimation.

est_tab: lavaan parameter table for estimation.

skewness: Skewness value(s) for the observed variables.

kurtosis: Kurtosis value(s) for the observed variables.

alpha: Significance criterion.

save_path: The directory (data) files were saved to.

results: Data frame containing the power analysis results

errors: A logical vector denoting failed Monte Carlo replications,

not_converged: A logical vector denoting non-converged Monte Carlo replications.

inadmissible A logical vector denoting Monte Carlo replications that resulted in negative variances or non-positive definite matrices.

The "session" element is a list containing information common to all experimental conditions, including

Psi: Residual variance-covariance matrix for the within-components.

reps: The number of Monte Carlo replication.

target_power: The desired power level.

save_path: The directory (data) files were saved to.

parameter: The parameter of primary interest.

constraints: The constraints imposed on the estimation model.

bounds: A logical denoting if bounded estimation was used for the latent variable variances in the model.

Details

A rationale for the power analysis strategy as implemented in this package can be found in Mulder (2022).

Data generation and model estimation

Data are generated using simulateData from the lavaan package. The within-components in the data generating model have variance of 1 such that the lagged effects as specified in Phi can be interpreted as standardized effects. This also implies that Phi is subject to the stationarity constraints of a VAR(1) model (i.e., eigenvalues smaller than 1). Based on Phi and wSigma, the residual variances-covariance for the within-components at wave 2 and later is computed (see compute_Psi).

Data are analyzed using lavaan from the lavaan package. The default estimator is maximum likelihood. The population parameter values are used as starting values. Optionally, the user can use bounded estimation, or impose constraints on parameters over time, as explained in the respective arguments.

Naming conventions for observed and latent variables

The observed variables in the RI-CLPM are given default names, namely capital letters in alphabetical order, with numbers denoting the measurement occasion. For example, for a bivariate RICLPM with 3 time points, we observe A1, A2, A3, B1, B2, and B3. Their within-components are denoted by wA1, wA2, ..., wB3, respectively. The between-components have RI_ prepended to the variable name, resulting in RI_A and RI_B.

Parameters are denoted using lavaan model syntax. For more information, see https://lavaan.ugent.be/tutorial/syntax1.html. For example, the random intercept variances are denoted by RI_A~~RI_A and RI_B~~RI_B, the cross-lagged effects at the first wave as wB2~wA1 and wA2~wB1, and the autoregressive effects as wA2~wA1 and wB2~wB1. Use names_powRICLPM to extract parameter names from the powRICLPM object.

Parallel processing using `furrr`

To speed up the analysis, power analysis for multiple experimental conditions can be executed in parallel. This has been implemented using furrr. By default the analysis is executed sequentially (i.e., single-core). Parallel execution (i.e., multicore) can be setup using plan, for example plan(multisession, workers = 4). For more information and options, see https://furrr.futureverse.org.

Progress bar using `progressr`

A progress bar displaying the status of the power analysis has been implemented using progressr. By default, a simple progress bar will be shown. For more information on how to control this progress bar and several other notification options (e.g., auditory notifications), see https://progressr.futureverse.org.

See also

Author

Jeroen D. Mulder j.d.mulder@uu.nl

Examples

# Example - Simulate power across range of sample sizes
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)

# Setup parallel computing (multicore, speeding up the analysis)
if (FALSE) {
library(furrr)
library(progressr)
plan(multisession, workers = 4)

# Run preliminary analysis (with limited "reps")
with_progress({
  out1 <- powRICLPM(
    target_power = 0.8,
    search_lower = 500,
    search_upper = 1000,
    search_step = 100,
    time_points = c(3, 4),
    ICC = c(0.4, 0.5, 0.6),
    RI_cor = 0.3,
    Phi = Phi,
    wSigma = wSigma,
    reps = 5,
    seed = 123456
 )
})
}