Create Mplus syntax for RI-CLPM power analysis

powRICLPM_Mplus() creates and saves syntax for performing a Monte Carlo power analysis for the random intercept cross-lagged panel model (RI-CLPM) in Mplus. It can create Mplus model syntax across multiple experimental conditions simultaneously. Conditions are defined in terms of sample size, number of time points, and proportion of between-unit variance (ICC). See "Details" for information on a) the naming conventions of parameters, and b) the various constraints that can be imposed on the model.

Usage

powRICLPM_Mplus(
  search_lower = NULL,
  search_upper = NULL,
  search_step = 20,
  sample_size = NULL,
  time_points,
  ICC,
  RI_cor,
  Phi,
  wSigma,
  reps = 1000,
  seed = NA,
  save_path = getwd(),
  constraints = "none"
)

Arguments

search_lower

A positive integer, denoting the lower bound of a range of sample sizes.

search_upper

A positive integer, denoting the upper bound of a range of sample sizes.

search_step

A positive integer, denoting an increment in sample size.

sample_size

(optional) An integer (vector), indicating specific sample sizes at which to evaluate power, rather than specifying a range using the search_ arguments.

time_points

An integer (vector) with elements at least larger than 3, indicating number of time points.

ICC

A double (vector), denoting the proportion of variance at the between-unit level.

RI_cor

A double, denoting the correlation between random intercepts.

Phi

A matrix, with standardized autoregressive effects (on the diagonal) and cross-lagged effects (off-diagonal) in the population. Columns represent predictors and rows represent outcomes.

wSigma

A correlation matrix for the within-unit components.

reps

A positive integer, denoting the number of Monte Carlo replications to be used during simulations.

seed

An `integer` of length 1. If multiple cores are used, a seed of length 1 will be used to generate a full L'Ecuyer-CMRG seed for all cores (see furrr_options).

save_path

Character string denoting a path leading to the folder where to save the Mplus syntax files to.

constraints

(optional) A character string, specifying the type of constraints that should be imposed on the estimation model (see "Details").

Details

Syntax generation

Mplus model syntax is created in multiple steps: First, the MODEL POPULATION command syntax is created in which parameters are constrained to population values. Second, the MODEL command syntax is created for model estimation. Optionally, syntax for constraints on the estimation model, in the MODEL CONSTRAINTS command, is created next. Ultimately, the parameter tables are combined to form character vectors containing the Mplus syntax to be exported (see "Details" of powRICLPM for more information on the constraints options).

Naming conventions

Details on the naming conventions can be found in the "Details" section of powRICLPM.

Examples

# Example - Create Mplus syntax to simulate power across range of sample sizes
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)

# Create and save Mplus model syntax
if (FALSE) {
powRICLPM_Mplus(
  search_lower = 500,
  search_upper = 1000,
  search_step = 100,
  time_points = c(3, 4),
  ICC = c(0.4, 0.5, 0.6),
  RI_cor = 0.3,
  Phi = Phi,
  wSigma = wSigma,
  reps = 10000,
  seed = 1234
)
}