Run Monte Carlo Simulation For Single Condition

run_condition() runs a Monte Carlo simulation for a single given condition. It generates data based on the pop_synt element in "object", and estimates an RI-CLPM using the est_synt element in "object". Data generation and model estimation are done using lavaan.

Usage

run_condition(object, p)

Arguments

object

A list with information for running a single simulation condition. See "Details" for an overview of the elements this object must contain.

p

A progressr object

Value

A list.

Details

Input: Elements in "object"

To successfully run a Monte Carlo simulation, the "object" arguments needs the following elements:

sample_size: The sample size.

time_points: The number of time points.

reps: Number of replications.

pop_synt: lavaan model syntax containing population values for data generation.

pop_tab: lavaan parameter table for data generation.

est_synt: lavaan model syntax for estimation.

est_tab: lavaan parameter table for estimation.

skewness: The skewness value(s) for the observed variables.

kurtosis: The kurtosis value(s) for the observed variables.

alpha: The significance criterion.

save_path: Folder to which simulated data are saved (optional).

Output

This function adds the following elements to "object":

results: A data frame containing the results (i.e., population values, bias, standard error of the estimate, coverage, power, etc.),

errors: A logical vector denoting failed Monte Carlo replications,

not_converged: A logical vector denoting non-converged Monte Carlo replications, and

inadmissible A logical vector denoting Monte Carlo replications that resulted in negative variances or non-positive definite matrices.