Summarize powRICLPM analysis

S3 method for class `powRICLPM`. summary.powRICLPM summarizes and outputs the setup and results of the powRICLPM analysis either generally, for a specific parameter (when parameter argument is specified), or for a specific parameter in a specific experimental condition (when the parameter, sample_size, time_points, and ICC arguments are specified). Optionally, it can display the parameter names contained in object.

Usage

# S3 method for powRICLPM
summary(
  object,
  ...,
  names = FALSE,
  parameter = NULL,
  sample_size = NULL,
  time_points = NULL,
  ICC = NULL
)

Arguments

object

A powRICLPM object.

...

Additional arguments not affecting the summary produced.

names

A logical denoting if parameter names contained in object should be printed.

parameter

A character string denoting the parameter of interest.

sample_size

An integer denoting the sample size of interest.

time_points

An integer denoting the number of time points of interest.

ICC

A numeric value denoting the proportion of variance at the between-unit level of interest.

Value

The powRICLPM object with the parameter element in object$session set to the value of the parameter argument.

Details

summary.powRICLPM sets the parameter element in session with the value of the parameter argument, and returns the object. Saving this object might be useful as it renders the parameter argument for plot.powRICLPM() and coef.powRICLPM() superfluous.

Parameter names

When simulating the power across conditions with a varying number of time points, there are also a different numbers of parameters across the experimental conditions. By specifying the argument names = TRUE, this function returns the parameter names from the condition with the smallest number of parameters. As such, the returned parameter names are valid for each experimental condition.

Examples

# Example - Same starting point as example from ?powRICLPM()
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)

# Setup parallel computing (multicore, speeding up the analysis)
if (FALSE) {
library(furrr)
plan(multisession, workers = 4)

# Run the analysis
out1 <- powRICLPM(
  target_power = 0.8,
  search_lower = 500,
  search_upper = 1000,
  search_step = 100,
  time_points = c(3, 4),
  ICC = c(0.4, 0.5, 0.6),
  RI_cor = 0.3,
  Phi = Phi,
  wSigma = wSigma,
  reps = 100,
  seed = 123456
)

# General results (i.e., setup of powRICLPM analysis)
summary(out)

# General results and recommendation for specific parameter
summary(out, parameter = "wB2~wA1")

# Set parameter element of object
out_wB2wA1 <- summary(out, parameter = "wB2~wA1")
}