Perform Power Analysis for the RI-CLPM

powRICLPM() is used to perform a Monte Carlo power analysis for the random intercept cross-lagged panel model (RICLPM). It has three main functionalities: 1) Recommend a sample size for a desired power level for a specific parameter (i.e., à priori analysis), 2) simulate the power for conditions with varying sample sizes and number of time points (i.e., post hoc analysis), and 3) create Mplus model syntax for performing power analysis in Mplus.

Usage

powRICLPM(
  target_power = NULL,
  parameter = NULL,
  search_lower = NULL,
  search_upper = NULL,
  search_step = NULL,
  sample_size = NULL,
  time_points,
  ICC,
  RI_cor,
  Phi,
  wSigma,
  reps,
  cores = NULL,
  seed = NULL,
  save_path = NA
)

Arguments

target_power

A numeric value between 0 and 1, denoting the desired power level for which to recommend a sample size. Use only for à priori analyses.

parameter

A character string denoting the parameter for which to compute a recommended sample size. Use only for à priori analysis. See "Details" for naming defaults for parameters used in this package.

search_lower

A positive integer, denoting the lower bound of the range of sample sizes for which to simulate the power. Use only for à priori analysis.

search_upper

A positive integer, denoting the upper bound of the range of sample sizes for which to simulate the power. Use only for à priori analysis.

search_step

A positive integer, denoting the increment in sample sizes. Use only for à priori analysis.

sample_size

An integer or integer vector with elements larger than 1, indicating the sample sizes at which to evaluate power. Use only for post hoc analysis.

time_points

An integer or integer vector with elements larger than 3, indicating number of time points.

ICC

A numeric value denoting the proportion of variance at the between-unit level.

RI_cor

A numeric value denoting the correlation between random intercepts.

Phi

A matrix of standardized autoregressive and cross-lagged effects in the population. Columns represent predictors and rows represent outcomes.

wSigma

A correlation matrix for the within-unit components.

reps

A positive integer denoting the number of Monte Carlo replications to be used during simulations.

cores

A positive integer denoting the number of cores to use during the Monte Carlo simulations. Executing the analysis using multiple cores can significantly reduce computation time.

seed

An integer of length 1. If multiple cores are used, a seed of length 1 will be used to generate a full L'Ecuyer-CMRG seed for all cores. For more information, see furrr_options.

save_path

A character string naming the directory to save any (data) files to.

Value

A list containing a "conditions" and "session" element. The "condition" element is again a list, where each element is itself a list containing elements needed to run a power analysis for a specific condition. This includes:

sample_size: The sample size.

time_points: The number of time points.

ICC: The proportion of between-unit variance.

reps: Number of replications.

pop_synt: lavaan model syntax containing population values for data generation.

pop_tab: lavaan parameter table for data generation.

est_synt: lavaan model syntax for estimation.

est_tab: lavaan parameter table for estimation.

save_path: The directory (data) files were saved to.

results: Data frame containing the power analysis results.

fails: Logical vector indicating failed MCMC reps.

The "session" element is a list containing information common to all conditions. For à priori analyses, this includes:

type: The type of power analysis performed.

target_power: The desired power level.

parameter: The parameter of interest.

search_lower: The lower bound of the set of candidate sample sizes.

search_upper: The upper bound of the set of candidate sample sizes.

search_step: The increment in sample sizes.

reps: The number of Monte Carlo replication.

For post hoc analyses, this includes:

type: The type of power analysis performed.

sample_size: The sample sizes of interest.

time_points: The number of time points of interest.

ICC: The proportion of between-unit variance.

reps: The number of Monte Carlo replication.

Details

Data generation

Data are generated from an RICLPM with a variance of 1 for the within-unit components, such that the lagged effects (specified in the Phi argument) can be interpreted as standardized effects. This implies that the process at the within-unit level must be stationary and that the Phi matrix is subject to the stationarity constraints of a VAR(1) model (i.e., eigenvalues smaller than 1). Based on the specified Phi and wSigma matrix, the residual variances and covariance between the within-unit components are computed (see compute_Psi) for wave 2 and later.

Naming conventions

The variables in the RICLPM are given default names, namely capital letters in alphabetical order, with numbers denoting the measurement occasion. For example, for a bivariate RICLPM with 3 time points, we have A1, A2, A3, B1, B2, and B3 as observed variables. Their within-components denoted by wA1, wA2, ..., wB3, respectively. The between components have RI_ prepended to the variable name, resulting in RI_A and RI_B.

Parameters are denoted using lavaan model syntax. For more information, see https://lavaan.ugent.be/tutorial/syntax1.html. For example, the random intercepts are denoted as RI_A~~RI_B, the cross-lagged effects at the first wave as wB2~wA1 and wA2~wB1, and the autoregressive effects as wA2~wA1 and wB2~wB1.

Data analysis

Data is generated and analyzed using lavaan. To speed up the process, power analysis for different conditions can be run in parallel (implemented using furrr).

See also

Author

Jeroen D. Mulder j.d.mulder@uu.nl

Examples

# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)

# Option 1 - À priori power analysis
output1 <- powRICLPM(target_power = 0.5,
                     parameter = "wB2~wA1",
                     search_lower = 500,
                     search_upper = 600,
                     search_step = 50,
                     time_points = 3,
                     ICC = 0.5,
                     RI_cor = 0.3,
                     Phi = Phi,
                     wSigma = wSigma,
                     reps = 30,
                     cores = 1,
                     seed = 123456)
#> 
#> A priori powRICLPM-analysis completed:
#>  - Monte Carlo replications:  30
#>  - Target power:  0.5
#>  - Preliminary recommended sample size:  500
#>  - Number of non-converged replications for recommended run:  0
#>  - Number of replications with inadmissible estimates for recommended run:  0
#> 
#> Sugested next steps:
#>  - Use powRICLPM_plot() to visualize preliminary results for entire range of sample sizes.
#>  - Rerun powRICLPM with the recommended sample size and large number of replications to validate preliminary results.

# Option 2 - Post hoc power analysis
output2 <- powRICLPM(sample_size = c(250, 350, 400),
                     time_points = c(3, 4),
                     ICC = 0.5,
                     RI_cor = 0.3,
                     Phi = Phi,
                     wSigma = wSigma,
                     reps = 100,
                     cores = 1,
                     seed = 123456)
#> 
#> Post hoc powRICLPM-analysis completed:
#>  - Monte Carlo replications:  100
#>  - Proportion random intercept variance: 0.5
#>  - Sample sizes (min. - max.): 250 - 400
#>  - Time points (min. - max.): 3 - 4
#> 
#> Suggestions:
#>  - Specify the `parameter` argument in the powRICLPM_summary() function to obtain detailed results for a specific parameter.
#>  - Use powRICLPM_plot() to visualize the power across all conditions.