Summarize setup and results from powRICLPM object

S3 method for class powRICLPM. summary.powRICLPM summarizes and outputs the setup and results of the powRICLPM analysis. Depending on the arguments that are set, summary.powRICLPM provides a different summary (see "Details").

Usage

# S3 method for powRICLPM
summary(
  object,
  ...,
  parameter = NULL,
  sample_size = NULL,
  time_points = NULL,
  ICC = NULL
)

Arguments

object

A powRICLPM object.

...

(don't use) Additional arguments not affecting the summary produced.

parameter

Character string of length denoting the parameter to visualize the results for.

sample_size

(optional) An integer, denoting the sample size of the experimental condition of interest.

time_points

(optional) An integer, denoting the number of time points of the experimental condition of interest.

ICC

(optional) A double, denoting the proportion of variance at the between-unit level of the experimental condition of interest.

Value

No return value, called for side effects.

Details

summary.powRICLPM provides a different summary of the powRICLPM object, depending on the additional arguments that are set:

When sample_size = ..., time_points = ..., and ICC = ... are set: Estimation information and results for all parameters of the experimental condition denoted by sample_size, time_points, and ICC.

When parameter = "..." is set: Estimation information and results for a specific parameter across all experimental conditions.

No additional arguments: Characteristics of the different experimental conditions are summarized, as well as session info (information that applies to each conditions, such the number of replications, etc.).

Examples

# \dontshow{
load(system.file("extdata", "out_preliminary.RData", package = "powRICLPM"))
# }

# Get setup of powRICLPM analysis and convergence issues
summary(out_preliminary)
#> powRICLPM (0.1.0.9001) simulated power for 16 experimental conditions:
#> 
#>  Sample size    Time points   ICC   Errors   Non-convergence    Inadmissible results
#> -------------  ------------  ----  -------  -----------------  ---------------------
#>      500                  3   0.4        0          0                              1
#>      600                  3   0.4        0          0                              0
#>      700                  3   0.4        0          0                              0
#>      800                  3   0.4        0          0                              1
#>      500                  4   0.4        0          0                              0
#>      600                  4   0.4        0          0                              0
#>      700                  4   0.4        0          0                              0
#>      800                  4   0.4        0          0                              0
#>      500                  3   0.6        0          0                              0
#>      600                  3   0.6        0          0                              0
#>      700                  3   0.6        0          0                              0
#>      800                  3   0.6        0          0                              0
#>      500                  4   0.6        0          0                              0
#>      600                  4   0.6        0          0                              0
#>      700                  4   0.6        0          0                              0
#>      800                  4   0.6        0          0                              0

# Performance measures for "wB2~wA1" parameter across experimental conditions
summary(out_preliminary, parameter = "wB2~wA1")
#> Results for wB2~wA1 across conditions:
#> 
#>    N    T   ICC   Err   NC   IR    PV     Avg      Min      SD   SE Avg     MSE   Accuracy   Coverage   Power
#> ----  ---  ----  ----  ---  ---  ----  ------  -------  ------  -------  ------  ---------  ---------  ------
#>  500    3   0.4     0    0    1   0.2   0.208    0.026   0.076    0.079   0.006      0.307       0.96    0.74
#>  600    3   0.4     0    0    0   0.2   0.205    0.004   0.071    0.073   0.005      0.286       0.96    0.82
#>  700    3   0.4     0    0    0   0.2   0.193   -0.005   0.063    0.069   0.004      0.271       0.96    0.85
#>  800    3   0.4     0    0    1   0.2   0.201    0.060   0.063    0.063   0.004      0.246       0.94    0.82
#>  500    4   0.4     0    0    0   0.2   0.191    0.078   0.055    0.062   0.003      0.242       0.98    0.87
#>  600    4   0.4     0    0    0   0.2   0.202    0.079   0.050    0.057   0.002      0.222       1.00    0.95
#>  700    4   0.4     0    0    0   0.2   0.206    0.072   0.061    0.053   0.004      0.206       0.91    0.96
#>  800    4   0.4     0    0    0   0.2   0.205    0.042   0.055    0.049   0.003      0.193       0.90    0.99
#>  500    3   0.6     0    0    0   0.2   0.195   -0.016   0.087    0.097   0.008      0.380       0.96    0.53
#>  600    3   0.6     0    0    0   0.2   0.188   -0.009   0.095    0.090   0.009      0.352       0.92    0.52
#>  700    3   0.6     0    0    0   0.2   0.201    0.007   0.088    0.083   0.008      0.325       0.94    0.68
#>  800    3   0.6     0    0    0   0.2   0.192    0.029   0.074    0.077   0.005      0.300       0.99    0.64
#>  500    4   0.6     0    0    0   0.2   0.203    0.036   0.068    0.071   0.005      0.278       0.95    0.77
#>  600    4   0.6     0    0    0   0.2   0.190    0.015   0.072    0.066   0.005      0.260       0.90    0.81
#>  700    4   0.6     0    0    0   0.2   0.201    0.019   0.065    0.060   0.004      0.236       0.96    0.92
#>  800    4   0.6     0    0    0   0.2   0.199    0.068   0.048    0.056   0.002      0.220       0.97    0.96

# Performance measures for all parameters, for specific experimental condition
summary(out_preliminary, sample_size = 600, time_points = 4, ICC = .4)
#> 
#> Results:
#> 
#> Parameter     Population value     Avg      Min      SD   SE Avg     MSE   Accuracy   Coverage  Power 
#> -----------  -----------------  ------  -------  ------  -------  ------  ---------  ---------  ------
#> RI_A~~RI_A               0.667   0.667    0.454   0.096    0.091   0.009      0.357       0.95  1.00  
#> RI_B~~RI_B               0.667   0.660    0.427   0.090    0.083   0.008      0.324       0.93  1.00  
#> RI_A~~RI_B               0.200   0.189    0.066   0.067    0.068   0.005      0.266       0.98  0.76  
#> wA2~wA1                  0.400   0.403    0.230   0.063    0.062   0.004      0.243       0.92  1.00  
#> wA2~wB1                  0.100   0.098   -0.024   0.056    0.055   0.003      0.215       0.94  0.47  
#> wB2~wA1                  0.200   0.202    0.079   0.050    0.057   0.002      0.222       1.00  0.95  
#> wB2~wB1                  0.300   0.292    0.132   0.061    0.060   0.004      0.237       0.95  1.00  
#> wA3~wA2                  0.400   0.395    0.219   0.063    0.062   0.004      0.244       0.95  1.00  
#> wA3~wB2                  0.100   0.104   -0.042   0.061    0.056   0.004      0.219       0.93  0.40  
#> wB3~wA2                  0.200   0.205    0.106   0.048    0.058   0.002      0.227       0.99  0.96  
#> wB3~wB2                  0.300   0.301    0.114   0.058    0.062   0.003      0.245       0.96  0.98  
#> wA4~wA3                  0.400   0.393    0.260   0.065    0.057   0.004      0.225       0.94  1.00  
#> wA4~wB3                  0.100   0.108   -0.015   0.056    0.052   0.003      0.205       0.92  0.56  
#> wB4~wA3                  0.200   0.199    0.037   0.054    0.055   0.003      0.217       0.97  0.95  
#> wB4~wB3                  0.300   0.296    0.139   0.059    0.057   0.003      0.224       0.94  1.00  
#> wA1~~wA1                 1.000   0.997    0.785   0.098    0.097   0.010      0.381       0.97  1.00  
#> wB1~~wB1                 1.000   0.994    0.763   0.084    0.091   0.007      0.356       0.96  1.00  
#> wA1~~wB1                 0.300   0.296    0.129   0.072    0.072   0.005      0.283       0.96  1.00  
#> wA2~~wA2                 0.806   0.800    0.650   0.058    0.061   0.003      0.238       0.94  1.00  
#> wA3~~wA3                 0.806   0.796    0.645   0.063    0.059   0.004      0.232       0.92  1.00  
#> wA4~~wA4                 0.806   0.803    0.659   0.057    0.056   0.003      0.219       0.93  1.00  
#> wB2~~wB2                 0.834   0.816    0.666   0.064    0.064   0.004      0.250       0.89  1.00  
#> wB3~~wB3                 0.834   0.829    0.682   0.067    0.063   0.004      0.246       0.94  1.00  
#> wB4~~wB4                 0.834   0.829    0.689   0.060    0.059   0.004      0.230       0.95  1.00  
#> wA2~~wB2                 0.148   0.146    0.033   0.044    0.044   0.002      0.174       0.93  0.94  
#> wA3~~wB3                 0.148   0.148    0.025   0.039    0.043   0.002      0.170       0.96  0.91  
#> wA4~~wB4                 0.148   0.147    0.061   0.041    0.041   0.002      0.160       0.95  0.96  
#> 
#> Estimation problems:
#> 
#> ----------------------  ---
#> Errors:                   0
#> Nonconvergence:           0
#> Inadmissible results:     0
#> ----------------------  ---
#> 
#> Condition info:
#> 
#> ------------------------  ------
#> Skewness:                      0
#> Kurtosis:                      0
#> Constraints:                none
#> Bounds:                    FALSE
#> Estimate ME:               FALSE
#> Significance criterion:     0.05
#> ------------------------  ------