Summarize setup and results from powRICLPM object

S3 method for class powRICLPM. summary.powRICLPM summarizes and outputs the setup and results of the powRICLPM analysis. Depending on the arguments that are set, summary.powRICLPM provides a different summary (see "Details").

Usage

# S3 method for powRICLPM
summary(
  object,
  ...,
  parameter = NULL,
  sample_size = NULL,
  time_points = NULL,
  ICC = NULL
)

Arguments

object

A powRICLPM object.

...

(don't use) Additional arguments not affecting the summary produced.

parameter

Character string of length denoting the parameter to visualize the results for.

sample_size

(optional) An integer, denoting the sample size of the experimental condition of interest.

time_points

(optional) An integer, denoting the number of time points of the experimental condition of interest.

ICC

(optional) A double, denoting the proportion of variance at the between-unit level of the experimental condition of interest.

Details

summary.powRICLPM provides a different summary of the powRICLPM object, depending on the additional arguments that are set:

When sample_size = ..., time_points = ..., and ICC = ... are set: Estimation information and results for all parameters of the experimental condition denoted by sample_size, time_points, and ICC.

When parameter = "..." is set: Estimation information and results for a specific parameter across all experimental conditions.

No additional arguments: Characteristics of the different experimental conditions are summarized, as well as session info (information that applies to each conditions, such the number of replications, etc.).

Examples

# Example - Same starting point as example from ?powRICLPM()
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)

# Setup parallel computing (multicore, speeding up the analysis)
if (FALSE) {
library(furrr)
plan(multisession, workers = 4)

# Run the analysis
out1 <- powRICLPM(
  target_power = 0.8,
  search_lower = 500,
  search_upper = 1000,
  search_step = 100,
  time_points = c(3, 4),
  ICC = c(0.4, 0.5, 0.6),
  RI_cor = 0.3,
  Phi = Phi,
  wSigma = wSigma,
  reps = 100,
  seed = 123456
)

# General results (i.e., setup of powRICLPM analysis)
summary(out)

# General results and recommendation for specific parameter
summary(out, parameter = "wB2~wA1")

# Summary for specific condition
summary(out, sample_size = 600, time_points = 4, ICC = .5)
}