Summarize Results from powRICLPM Object

S3 method for class powRICLPM. summary.powRICLPM summarizes the setup and results of the powRICLPM analysis. Depending on the arguments that are set, summary.powRICLPM provides a different summary (see "Details").

Usage

# S3 method for class 'powRICLPM'
summary(
  object,
  ...,
  parameter = NULL,
  sample_size = NULL,
  time_points = NULL,
  ICC = NULL,
  reliability = NULL
)

Arguments

object

A powRICLPM object.

...

(don't use)

parameter

Character string of length 1 denoting the parameter to visualize the results for.

sample_size

(optional) An integer, denoting the sample size of the experimental condition of interest.

time_points

(optional) An integer, denoting the number of time points of the experimental condition of interest.

ICC

(optional) A double, denoting the proportion of variance at the between-unit level of the experimental condition of interest.

reliability

(optional) An integer, denoting the reliability of the indicators of the experimental condition of interest.

Value

No return value, called for side effects.

Details

summary.powRICLPM provides a different summary of the powRICLPM object, depending on the additional arguments that are set:

When sample_size = ..., time_points = ..., ICC = ..., and reliability are set: Estimation information and results for all parameters across experimental conditions.

When parameter = "..." is set: Estimation information and results for a specific parameter across all experimental conditions.

No additional arguments: Characteristics of the different experimental conditions are summarized, as well as session info (information that applies to all conditions, such the number of replications, etc.).

Examples

# Get setup of powRICLPM analysis and convergence issues
summary(out_preliminary)
#> powRICLPM (0.1.1.9000) simulated power for 16 experimental conditions.
#> 
#> 
#> Table: SUMMARY OF ANALYSIS PER EXPERIMENTAL CONDITION
#> 
#>  Sample size   Time points   ICC   Reliability   Error   Not converged   Inadmissible
#> ------------  ------------  ----  ------------  ------  --------------  -------------
#>          500             4   0.3           1.0       0               0              0
#>          700             4   0.3           1.0       0               0              0
#>          500             5   0.3           1.0       0               0              0
#>          700             5   0.3           1.0       0               0              0
#>          500             4   0.5           1.0       0               0              0
#>          700             4   0.5           1.0       0               0              0
#>          500             5   0.5           1.0       0               0              0
#>          700             5   0.5           1.0       0               0              0
#>          500             4   0.3           0.8       0               0              0
#>          700             4   0.3           0.8       0               0              0
#>          500             5   0.3           0.8       0               0              0
#>          700             5   0.3           0.8       0               0              0
#>          500             4   0.5           0.8       0               0              0
#>          700             4   0.5           0.8       0               0              0
#>          500             5   0.5           0.8       0               0              0
#>          700             5   0.5           0.8       0               0              0

# Performance measures for "wB2~wA1" parameter across experimental conditions
summary(out_preliminary, parameter = "wB2~wA1")
#> 
#> 
#> Table: SIMULATION RESULTS FOR wB2~wA1
#> 
#>  Sample size   Time points   ICC   Reliability   Population     Avg      Min       SD   SE Avg     MSE   Accuracy   Cover   Power   Error   Not converged   Inadmissible   NA
#> ------------  ------------  ----  ------------  -----------  ------  -------  -------  -------  ------  ---------  ------  ------  ------  --------------  -------------  ---
#>          500             4   0.3           0.8          0.1   0.092   -0.008   -0.034    0.051   0.056      0.003   0.218    0.96    0.34               0              0    0
#>          500             4   0.3           1.0          0.1   0.101    0.001   -0.037    0.061   0.057      0.004   0.225    0.90    0.44               0              0    0
#>          500             4   0.5           0.8          0.1   0.078   -0.022   -0.072    0.058   0.061      0.004   0.238    0.91    0.23               0              0    0
#>          500             4   0.5           1.0          0.1   0.093   -0.007   -0.047    0.063   0.065      0.004   0.254    0.95    0.32               0              0    0
#>          500             5   0.3           0.8          0.1   0.091   -0.009   -0.026    0.046   0.053      0.002   0.208    0.98    0.40               0              0    0
#>          500             5   0.3           1.0          0.1   0.102    0.002   -0.036    0.057   0.054      0.003   0.213    0.94    0.50               0              0    0
#>          500             5   0.5           0.8          0.1   0.084   -0.016   -0.097    0.055   0.057      0.003   0.222    0.93    0.37               0              0    0
#>          500             5   0.5           1.0          0.1   0.093   -0.007   -0.039    0.062   0.059      0.004   0.233    0.95    0.34               0              0    0
#>          700             4   0.3           0.8          0.1   0.072   -0.028   -0.023    0.043   0.048      0.003   0.186    0.97    0.33               0              0    0
#>          700             4   0.3           1.0          0.1   0.105    0.005    0.011    0.048   0.049      0.002   0.192    0.98    0.58               0              0    0
#>          700             4   0.5           0.8          0.1   0.074   -0.026   -0.036    0.052   0.052      0.003   0.203    0.94    0.33               0              0    0
#>          700             4   0.5           1.0          0.1   0.102    0.002   -0.054    0.060   0.054      0.004   0.214    0.90    0.49               0              0    0
#>          700             5   0.3           0.8          0.1   0.083   -0.017   -0.025    0.046   0.045      0.002   0.176    0.92    0.39               0              0    0
#>          700             5   0.3           1.0          0.1   0.098   -0.002    0.018    0.041   0.046      0.002   0.179    0.98    0.59               0              0    0
#>          700             5   0.5           0.8          0.1   0.077   -0.023   -0.051    0.049   0.047      0.003   0.185    0.90    0.40               0              0    0
#>          700             5   0.5           1.0          0.1   0.099   -0.001   -0.030    0.051   0.049      0.003   0.193    0.92    0.57               0              0    0

# Performance measures for all parameters, for specific experimental condition
summary(out_preliminary, sample_size = 700, time_points = 4, ICC = .3, reliability = 1)
#> 
#> 
#> 
#> Table: SUMMARY OF ANALYSIS
#> 
#>                         Number of replications 
#> ----------------------  -----------------------
#> Requested:              100                    
#> Completed:              100                    
#> Convergence issues:     0                      
#> Inadmissible results:   0                      
#> 
#> 
#> 
#> Table: SUMMARY OF SIMULATION CONDITION
#> 
#>                                Value 
#> -----------------------------  ------
#> Skewness:                      0     
#> Kurtosis:                      0     
#> Constraints:                   none  
#> Bounds:                        FALSE 
#> Estimated measurement error:   FALSE 
#> Significance criterion:        0.05  
#> 
#> 
#> 
#> Table: SIMULATION RESULTS
#> 
#>              Population      Avg      Min       SD   SE Avg     MSE   Accuracy   Cover   Power  NA   
#> -----------  -----------  ------  -------  -------  -------  ------  ---------  ------  ------  -----
#> RI_A~~RI_A   0.429         0.426   -0.003    0.261    0.058   0.055      0.003   0.217    0.97  1.00 
#> RI_B~~RI_B   0.429         0.425   -0.003    0.269    0.071   0.062      0.005   0.243    0.89  1.00 
#> RI_A~~RI_B   0.129         0.125   -0.004    0.012    0.055   0.045      0.003   0.178    0.89  0.69 
#> wA2~wA1      0.200         0.200    0.000    0.051    0.060   0.054      0.004   0.210    0.94  0.94 
#> wA2~wB1      0.150         0.150    0.000    0.017    0.053   0.051      0.003   0.200    0.93  0.85 
#> wB2~wA1      0.100         0.105    0.005    0.011    0.048   0.049      0.002   0.192    0.98  0.58 
#> wB2~wB1      0.300         0.299   -0.001    0.116    0.056   0.054      0.003   0.210    0.94  0.99 
#> wA3~wA2      0.200         0.197   -0.003    0.013    0.055   0.054      0.003   0.213    0.97  0.95 
#> wA3~wB2      0.150         0.160    0.010    0.020    0.054   0.052      0.003   0.205    0.93  0.85 
#> wB3~wA2      0.100         0.101    0.001   -0.030    0.051   0.050      0.003   0.195    0.96  0.56 
#> wB3~wB2      0.300         0.295   -0.005    0.139    0.063   0.056      0.004   0.218    0.92  1.00 
#> wA4~wA3      0.200         0.196   -0.004    0.083    0.050   0.051      0.003   0.202    0.95  0.97 
#> wA4~wB3      0.150         0.153    0.003    0.035    0.045   0.050      0.002   0.195    0.97  0.90 
#> wB4~wA3      0.100         0.103    0.003   -0.018    0.048   0.047      0.002   0.185    0.94  0.56 
#> wB4~wB3      0.300         0.302    0.002    0.176    0.053   0.051      0.003   0.199    0.92  1.00 
#> wA1~~wA1     1.000         0.996   -0.004    0.844    0.081   0.074      0.006   0.288    0.90  1.00 
#> wB1~~wB1     1.000         1.006    0.006    0.826    0.081   0.078      0.006   0.306    0.95  1.00 
#> wA1~~wB1     0.300         0.302    0.002    0.174    0.063   0.057      0.004   0.224    0.92  1.00 
#> wA2~~wA2     0.919         0.923    0.003    0.781    0.066   0.063      0.004   0.248    0.93  1.00 
#> wA3~~wA3     0.919         0.918   -0.002    0.764    0.059   0.062      0.003   0.244    0.95  1.00 
#> wA4~~wA4     0.919         0.915   -0.004    0.783    0.061   0.059      0.004   0.230    0.92  1.00 
#> wB2~~wB2     0.882         0.872   -0.010    0.722    0.062   0.059      0.004   0.232    0.88  1.00 
#> wB3~~wB3     0.882         0.892    0.010    0.763    0.057   0.059      0.003   0.233    0.98  1.00 
#> wB4~~wB4     0.882         0.869   -0.013    0.730    0.057   0.055      0.003   0.215    0.95  1.00 
#> wA2~~wB2     0.212         0.214    0.001    0.104    0.046   0.044      0.002   0.173    0.95  1.00 
#> wA3~~wB3     0.212         0.214    0.001    0.123    0.043   0.044      0.002   0.172    0.97  1.00 
#> wA4~~wB4     0.212         0.213    0.001    0.090    0.041   0.041      0.002   0.161    0.95  1.00